refactoring + update
update of the class User

refactoring of the class UserManager

refactoring of the Database class

first update of "stake" command

+++ utils/AdminSettings/admin_settings.h
@@ -19,7 +19,7 @@ public:
 
     static std::string getValueByKey(const std::string key);
     
-    static double getMoney();
+    static int getMoney();
 
     static int dailyBets();
